fix(874): fix bug with tapbar for android on filter screen
