+++ src/navigation/navigators/MainNavigator.tsx
@@ -6,6 +6,8 @@ import {
   ImagesGalleryScreen,
   ObjectDetailsAddInfoScreen,
   ObjectDetailsShareExperienceScreen,
+  FiltersScreen,
+  SettlementsScreen,
 } from '../../screens';
 
 import {TabNavigator} from './TabNavigator';
@@ -13,14 +15,18 @@ import {MainNavigatorParamsList} from 'core/types';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {defaultTransition} from '../transition';
 import {AuthNavigator} from './AuthNavigator';
+import {useTranslation, useColorScheme} from 'core/hooks';
 import {useAndroidNavbarStyle, useScreenOptions} from 'navigation/hooks';
 import {useSelector} from 'react-redux';
 import {selectObjectShareExperienceData} from 'core/selectors';
+import {COLORS} from 'assets';
 
 const Stack = createNativeStackNavigator<MainNavigatorParamsList>();
 
 export function MainNavigator() {
   useAndroidNavbarStyle();
+  const {t: tFiters} = useTranslation('filters');
+  const colorScheme = useColorScheme();
 
   const screenOptions = useScreenOptions({
     animation: defaultTransition,
@@ -50,7 +56,36 @@ export function MainNavigator() {
         component={ObjectDetailsMapScreen}
         options={{headerShown: false}}
       />
-
+      <Stack.Group
+        screenOptions={() => ({
+          headerStyle: {
+            backgroundColor:
+              colorScheme === 'light'
+                ? COLORS.light.background.primary
+                : COLORS.dark.background.primary,
+          },
+          presentation: 'modal',
+        })}>
+        <Stack.Screen
+          name="Filter"
+          component={FiltersScreen}
+          options={props => ({
+            ...FiltersScreen.screenOptions(props),
+          })}
+        />
+        <Stack.Screen
+          name="Settlements"
+          component={SettlementsScreen}
+          options={props => ({
+            ...SettlementsScreen.screenOptions(props),
+            title: tFiters('settlements.title'),
+            headerTintColor:
+              colorScheme === 'light'
+                ? COLORS.light.text.primary
+                : COLORS.dark.text.primary,
+          })}
+        />
+      </Stack.Group>
       <Stack.Screen
         name="PageNotFoundErrorScreen"
         component={PageNotFoundErrorScreen}
