+++ src/screens/Filters/types.ts
@@ -2,21 +2,17 @@ import {StackNavigationProp} from '@react-navigation/stack';
 import {NativeStackNavigationOptions} from '@react-navigation/native-stack';
 import {BottomTabNavigationProp} from '@react-navigation/bottom-tabs';
 import {RouteProp, CompositeNavigationProp} from '@react-navigation/native';
-import {
-  HomeNavigatorParamsList,
-  TabNavigatorParamsList,
-  MainNavigatorParamsList,
-} from 'core/types';
+import {TabNavigatorParamsList, MainNavigatorParamsList} from 'core/types';
 
 export type HomeScreenNavigationProps = CompositeNavigationProp<
-  StackNavigationProp<HomeNavigatorParamsList, 'Filter'>,
+  StackNavigationProp<MainNavigatorParamsList, 'Filter'>,
   CompositeNavigationProp<
     BottomTabNavigationProp<TabNavigatorParamsList>,
     StackNavigationProp<MainNavigatorParamsList>
   >
 >;
 
-export type HomeScreenRouteProps = RouteProp<HomeNavigatorParamsList, 'Filter'>;
+export type HomeScreenRouteProps = RouteProp<MainNavigatorParamsList, 'Filter'>;
 
 export interface IProps {
   navigation: HomeScreenNavigationProps;
