+++ src/screens/Settlements/types.ts
@@ -2,14 +2,10 @@ import {StackNavigationProp} from '@react-navigation/stack';
 import {NativeStackNavigationOptions} from '@react-navigation/native-stack';
 import {BottomTabNavigationProp} from '@react-navigation/bottom-tabs';
 import {RouteProp, CompositeNavigationProp} from '@react-navigation/native';
-import {
-  HomeNavigatorParamsList,
-  TabNavigatorParamsList,
-  MainNavigatorParamsList,
-} from 'core/types';
+import {TabNavigatorParamsList, MainNavigatorParamsList} from 'core/types';
 
 export type HomeScreenNavigationProps = CompositeNavigationProp<
-  StackNavigationProp<HomeNavigatorParamsList, 'Settlements'>,
+  StackNavigationProp<MainNavigatorParamsList, 'Settlements'>,
   CompositeNavigationProp<
     BottomTabNavigationProp<TabNavigatorParamsList>,
     StackNavigationProp<MainNavigatorParamsList>
@@ -17,7 +13,7 @@ export type HomeScreenNavigationProps = CompositeNavigationProp<
 >;
 
 export type SettlementsScreenRouteProps = RouteProp<
-  HomeNavigatorParamsList,
+  MainNavigatorParamsList,
   'Settlements'
 >;
